% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_history.R
\name{download_history}
\alias{download_history}
\title{Downloads history}
\usage{
download_history()
}
\value{
A `data.frame` with columns: package, version, snapshot_time, folder,
subfolder, submission_time.
}
\description{
Downloads history of packages on the submission queue as recorded on github branch.
}
\details{
For some periods github actions recording the data didn't run,
so there are some periods with missing data.
}
