% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from a submission}
\usage{
get_metadata(snapshot_row, dir = tempdir(check = TRUE))
}
\arguments{
\item{snapshot_row}{a row of \code{take_snapshot()} output}

\item{dir}{folder where to download and untar information.}
}
\value{
tibble
}
\description{
Get metadata from a submission
}
