% Generated by roxygen2: do not edit by hand
% Please edit documentation in r/take-snapshot.R
\name{take_snapshot}
\alias{take_snapshot}
\title{Take Snapshot of CRAN incoming folder}
\usage{
take_snapshot()
}
\value{
A data.frame, one line per submission.
}
\description{
Take Snapshot of CRAN incoming folder
}
